refactor - travelcard 폴더 구조 components / travel로 변경
